# Conflicts:
#	lambdas/enums/lambda_error.py
#	lambdas/models/document_reference.py
#	lambdas/services/base/dynamo_service.py
#	lambdas/tests/unit/services/test_upload_document_reference_service.py

+++ lambdas/services/base/s3_service.py
@@ -124,8 +124,11 @@ class S3Service:
             StorageClass="INTELLIGENT_TIERING",
         )
 
-    def delete_object(self, s3_bucket_name: str, file_key: str):
-        return self.client.delete_object(Bucket=s3_bucket_name, Key=file_key)
+    def delete_object(self, s3_bucket_name: str, file_key: str, version_id: str | None = None):
+        if version_id is None:
+            return self.client.delete_object(Bucket=s3_bucket_name, Key=file_key)
+        
+        return self.client.delete_object(Bucket=s3_bucket_name, Key=file_key, VersionId=version_id)
 
     def create_object_tag(
         self, s3_bucket_name: str, file_key: str, tag_key: str, tag_value: str
